size_t is not for file size

+++ include/ruby/internal/intern/file.h
@@ -187,7 +187,7 @@ RBIMPL_ATTR_PURE()
  */
 int rb_is_absolute_path(const char *path);
 
-size_t rb_file_size(VALUE file);
+off_t rb_file_size(VALUE file);
 
 RBIMPL_SYMBOL_EXPORT_END()
